requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
tensorflow>=2.9.0
nltk>=3.7
spacy>=3.4.0
flask>=2.2.0
psycopg2-binary>=2.9.3
pymongo>=4.2.0
python-dotenv>=0.20.0
pyjwt>=2.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
matplotlib>=3.5.0
seaborn>=0.12.0
transformers>=4.21.0 